add missing mount

+++ destroy-gentoo-base-container.sh
@@ -5,6 +5,11 @@ CONTAINER_NAME=owncloud-app
 
 lxc-stop -n ${CONTAINER_NAME}
 # NOT needed #lxc-destroy -n ${CONTAINER_NAME}
-umount /var/lib/lxc/${CONTAINER_NAME}
-rmdir /var/lib/lxc/${CONTAINER_NAME}
+
+mountpoint -q /var/lib/lxc/${CONTAINER_NAME}
+if [ $? -eq 0 ]; then
+	umount /var/lib/lxc/${CONTAINER_NAME}
+	rmdir /var/lib/lxc/${CONTAINER_NAME}
+fi
+mountpoint -q /mnt/full-root/ || mount /mnt/full-root
 btrfs sub delete /mnt/full-root/vols/${CONTAINER_NAME}
